back_end/core-master/requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
pytz==2020.5
sqlparse==0.4.1
psycopg2-binary==2.8.6
django-cors-headers==3.7.0
Pillow==8.1.2
drf-nested-routers==0.91
drf-yasg==1.20.0